+++ config/environment/test.php
@@ -9,13 +9,17 @@ return array(
 
     // Disable logging
     'Psr\Log\LoggerInterface' => \DI\decorate(function ($previous, ContainerInterface $c) {
-        $enableLogging = $c->get('ini.tests.enable_logging') == 1;
+        $enableLogging = $c->get('ini.tests.enable_logging') == 1 || !empty(getenv('MATOMO_TESTS_ENABLE_LOGGING'));
         if ($enableLogging) {
             return $previous;
         } else {
             return $c->get(\Psr\Log\NullLogger::class);
         }
     }),
+
+    'log.handlers' => [
+        \DI\get('Piwik\Plugins\Monolog\Handler\FileHandler'),
+    ],
 
     'Piwik\Cache\Backend' => function () {
         return \Piwik\Cache::buildBackend('file');